Refactor&Fix&Add: handler 분리, foreground 상태 대응, controller 에러 해결
