+++ android/app/src/main/AndroidManifest.xml
@@ -30,6 +30,9 @@
               android:name="io.flutter.embedding.android.NormalTheme"
               android:resource="@style/NormalTheme"
             />
+            <meta-data
+  android:name="flutterEmbedding"
+  android:value="2" />
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER"/>
